+++ API/src/main/java/fr/maxlego08/essentials/api/commands/CommandManager.java
@@ -2,6 +2,7 @@ package fr.maxlego08.essentials.api.commands;
 
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.TabCompleter;
+import org.bukkit.event.Listener;
 import org.bukkit.plugin.Plugin;
 
 import java.util.List;
@@ -9,7 +10,7 @@ import java.util.List;
 /**
  * Represents a manager for handling plugin commands.
  */
-public interface CommandManager extends CommandExecutor, TabCompleter {
+public interface CommandManager extends CommandExecutor, TabCompleter, Listener {
 
     /**
      * Registers a command with the command manager.
@@ -50,8 +51,16 @@ public interface CommandManager extends CommandExecutor, TabCompleter {
      */
     List<EssentialsCommand> getCommands();
 
+    /**
+     * Gets a list of all commands registered with the command manager, sorted by their main command.
+     *
+     * @return A list of all registered commands, sorted by their main command.
+     */
     List<EssentialsCommand> getSortCommands();
 
+    /**
+     * Saves all commands to the command configuration.
+     */
     void saveCommands();
 }