+++ src/main/java/fr/maxlego08/essentials/economy/EconomyModule.java
@@ -17,6 +17,7 @@ import fr.maxlego08.essentials.api.economy.PriceFormat;
 import fr.maxlego08.essentials.api.economy.UserBaltop;
 import fr.maxlego08.essentials.api.event.events.economy.EconomyBaltopUpdateEvent;
 import fr.maxlego08.essentials.api.event.events.user.UserFirstJoinEvent;
+import fr.maxlego08.essentials.api.event.events.user.UserJoinEvent;
 import fr.maxlego08.essentials.api.event.events.user.UserQuitEvent;
 import fr.maxlego08.essentials.api.messages.Message;
 import fr.maxlego08.essentials.api.storage.IStorage;
@@ -46,6 +47,8 @@ import java.util.function.Consumer;
 public class EconomyModule extends ZModule implements EconomyManager {
 
     @NonLoadable
+    public static final String NO_REASON = "No reason";
+    @NonLoadable
     private final List<Economy> economies = new ArrayList<>();
     private final List<NumberMultiplicationFormat> numberFormatSellMultiplication = new ArrayList<>();
     private final Map<Economy, Baltop> baltops = new HashMap<>();
@@ -62,6 +65,8 @@ public class EconomyModule extends ZModule implements EconomyManager {
     private long baltopRefreshSeconds;
     private String baltopPlaceholderUserEmpty;
     private String baltopMessageEconomy;
+    private String payWithdrawReason;
+    private String payDepositReason;
     private int baltopMessageAmount;
     private BaltopDisplay baltopDisplay;
     private WrappedTask baltopTask;
@@ -168,19 +173,46 @@ public class EconomyModule extends ZModule implements EconomyManager {
 
     @Override
     public boolean deposit(UUID uniqueId, Economy economy, BigDecimal amount) {
-        perform(uniqueId, user -> user.deposit(economy, amount));
-        return true;
-    }
-
-    @Override
-    public boolean withdraw(UUID uniqueId, Economy economy, BigDecimal amount) {
-        perform(uniqueId, user -> user.withdraw(economy, amount));
-        return true;
+        return deposit(uniqueId, economy, amount, NO_REASON);
     }
 
     @Override
     public boolean set(UUID uniqueId, Economy economy, BigDecimal amount) {
-        perform(uniqueId, user -> user.set(economy, amount));
+        return set(uniqueId, economy, amount, NO_REASON);
+    }
+
+    @Override
+    public boolean withdraw(UUID uniqueId, Economy economy, BigDecimal amount) {
+        return withdraw(uniqueId, economy, amount, NO_REASON);
+    }
+
+    @Override
+    public boolean deposit(UUID uniqueId, Economy economy, BigDecimal amount, String reason) {
+        perform(uniqueId, user -> {
+            user.deposit(economy, amount, reason);
+            var offlineEconomy = getOfflineEconomy(uniqueId);
+            offlineEconomy.deposit(economy.getName(), amount);
+        });
+        return true;
+    }
+
+    @Override
+    public boolean withdraw(UUID uniqueId, Economy economy, BigDecimal amount, String reason) {
+        perform(uniqueId, user -> {
+            user.withdraw(economy, amount, reason);
+            var offlineEconomy = getOfflineEconomy(uniqueId);
+            offlineEconomy.withdraw(economy.getName(), amount);
+        });
+        return true;
+    }
+
+    @Override
+    public boolean set(UUID uniqueId, Economy economy, BigDecimal amount, String reason) {
+        perform(uniqueId, user -> {
+            user.set(economy, amount, reason);
+            var offlineEconomy = getOfflineEconomy(uniqueId);
+            offlineEconomy.set(economy.getName(), amount);
+        });
         return true;
     }
 
@@ -265,8 +297,8 @@ public class EconomyModule extends ZModule implements EconomyManager {
     @Override
     public void pay(UUID fromUuid, String fromName, UUID toUuid, String toName, Economy economy, BigDecimal amount) {
 
-        perform(fromUuid, user -> user.withdraw(toUuid, economy, amount));
-        perform(toUuid, user -> user.deposit(fromUuid, economy, amount));
+        perform(fromUuid, user -> user.withdraw(toUuid, economy, amount, this.payWithdrawReason.replace("%player%", toName)));
+        perform(toUuid, user -> user.deposit(fromUuid, economy, amount, this.payDepositReason.replace("%player%", fromName)));
 
         message(fromUuid, Message.COMMAND_PAY_SENDER, "%amount%", this.format(economy, amount), "%player%", toName);
         message(toUuid, Message.COMMAND_PAY_RECEIVER, "%amount%", this.format(economy, amount), "%player%", fromName);
@@ -383,6 +415,11 @@ public class EconomyModule extends ZModule implements EconomyManager {
         User user = event.getUser();
         OfflineEconomy offlineEconomy = new ZOfflineEconomy(user.getBalances());
         this.offlinePlayers.put(user.getUniqueId(), offlineEconomy);
+    }
+
+    @EventHandler
+    public void onConnect(UserJoinEvent event) {
+        this.offlinePlayers.remove(event.getUser().getUniqueId());
     }
 
 }