+++ changelog.md
@@ -1,5 +1,25 @@
 # Unreleased
 
+# 1.0.1.5
+
+- Added ``/clearinventory [<player>]`` [#101](https://github.com/Maxlego08/zEssentials/issues/124)
+- Improve economy module with offline players
+- Fixed teleport command with relative coordinates [#142](https://github.com/Maxlego08/zEssentials/issues/142)
+- Added a method in the API to retrieve player’s transaction history
+- Added a reason for each transaction made by the player
+- Fixed sql port [#144](https://github.com/Maxlego08/zEssentials/issues/144)
+- Fixed rtp with folia [#138](https://github.com/Maxlego08/zEssentials/issues/138)
+- Fixed kit permission
+- Fixed method ``stringToDuration`` [#143](https://github.com/Maxlego08/zEssentials/pull/143)
+- Fixed the cooldown command for commands that don’t come from
+  zEssentials [#137](https://github.com/Maxlego08/zEssentials/pull/137)
+- Fixed teleport request [#134](https://github.com/Maxlego08/zEssentials/pull/134)
+- Added ``/suicide`` [#135](https://github.com/Maxlego08/zEssentials/pull/135)
+- Move ``commands.md``, `placeholders.md` and `permissions.md` in docs folder
+- Added ``/eco give <economy> <player> <min amount> <max amount>``
+  command [#120](https://github.com/Maxlego08/zEssentials/pull/120)
+- Fixed teleport command if player doesn't exit [#112](https://github.com/Maxlego08/zEssentials/pull/112)
+
 # 1.0.1.4
 
 - Fixed auto update task for hologram module
@@ -17,7 +37,8 @@
 - Added command ``/fly info``
 - Added command ``/ess delete-world <world>``, allows you to delete data related to a world
 - Fixed delete home sql request [#119](https://github.com/Maxlego08/zEssentials/issues/119)
-- Added permission ``essentials.fly.safelogin``, Players with this permission will automatically enter fly mode upon logging in if they are suspended in the air. [#117](https://github.com/Maxlego08/zEssentials/issues/117)
+- Added permission ``essentials.fly.safelogin``, Players with this permission will automatically enter fly mode upon
+  logging in if they are suspended in the air. [#117](https://github.com/Maxlego08/zEssentials/issues/117)
 - If the player does not have the `essentials.speed` permission, the walk and fly speed will be reset to default values
 - Tab completion for editing hologram and itemrename [#116](https://github.com/Maxlego08/zEssentials/issues/116)
 
@@ -46,7 +67,8 @@
 - Fixed message when the player leave server
 - Added permission ``essentials.back.death``, the player must have this permission to return to the place of his death
 - Fixed teleport task when player is offline [#92](https://github.com/Maxlego08/zEssentials/issues/92)
-- Fixed command ``/heal <player>`` and `/feed <player>` if you use it in the console [#90](https://github.com/Maxlego08/zEssentials/issues/90)
+- Fixed command ``/heal <player>`` and `/feed <player>` if you use it in the
+  console [#90](https://github.com/Maxlego08/zEssentials/issues/90)
 - Added two types of home usage MAX and STACK [#84](https://github.com/Maxlego08/zEssentials/issues/84)
 - Fixed home delete inventory [#89](https://github.com/Maxlego08/zEssentials/issues/89)
 - Fixed item display when you try to display an empty item
@@ -58,7 +80,7 @@
 # 1.0.1.0
 
 - Fixed command cooldown if permission was not set
-- Fixed the system of economy that did not work with offline players 
+- Fixed the system of economy that did not work with offline players
 
 # 1.0.0.9
 
@@ -68,15 +90,20 @@
 - Fixed an SQL query for updating cooldown with SQLITE [#74](https://github.com/Maxlego08/zEssentials/issues/74)
 - Fixed creating homes that executed an SQL query for no reason
 - Added command ``/freeze <player>``
-- Added a command to confirm the overwrite of an already existing home. You can disable this option in `modules/homes/config.yml`
+- Added a command to confirm the overwrite of an already existing home. You can disable this option in
+  `modules/homes/config.yml`
 - Added a command to confirm the deletion of an home. You can disable this option in `modules/homes/config.yml`
-- Fixed commands ``/sethome <player>:<home name>`` and ``/delhome <player>:<home name>`` which did not work if the player was online.
-- Fixed command ``/wtp`` with folia 
+- Fixed commands ``/sethome <player>:<home name>`` and ``/delhome <player>:<home name>`` which did not work if the
+  player was online.
+- Fixed command ``/wtp`` with folia
 
 # 1.0.0.8
 
-- Added error exception when you try to load a home if the world doesn't exist [#67](https://github.com/Maxlego08/zEssentials/issues/67)
-- Implementation of the method ``boolean hasMoney(OfflinePlayer player, Economy economy, BigDecimal amount)`` and `BigDecimal getBalance(OfflinePlayer player, Economy economy)` in `EconomyModule` [#66](https://github.com/Maxlego08/zEssentials/issues/66)
+- Added error exception when you try to load a home if the world doesn't
+  exist [#67](https://github.com/Maxlego08/zEssentials/issues/67)
+- Implementation of the method ``boolean hasMoney(OfflinePlayer player, Economy economy, BigDecimal amount)`` and
+  `BigDecimal getBalance(OfflinePlayer player, Economy economy)` in
+  `EconomyModule` [#66](https://github.com/Maxlego08/zEssentials/issues/66)
 - Fix sarah migration
 - Fixed invsee command [#72](https://github.com/Maxlego08/zEssentials/issues/72)
 
@@ -90,13 +117,15 @@
 
 # 1.0.0.6
 
-- Added the command ``/showitem <code>``, Allows you to see the item that the player has in his hand. This command is used with the chat placeholder `[item]` [#43](https://github.com/Maxlego08/zEssentials/issues/43)
+- Added the command ``/showitem <code>``, Allows you to see the item that the player has in his hand. This command is
+  used with the chat placeholder `[item]` [#43](https://github.com/Maxlego08/zEssentials/issues/43)
 - Added the command ``/money <player>``, Shows the money of other players.
 - Fixed messages [#39](https://github.com/Maxlego08/zEssentials/issues/39)
 - Fixed default configuration for economy [#38](https://github.com/Maxlego08/zEssentials/issues/38)
 - Fixed bug with ``/tp`` command [#37](https://github.com/Maxlego08/zEssentials/issues/37)
 - Fixed nightvision messages
 - Add the feature to manage vaults slots with permissions
+
 ```yaml
 vault-permissions:
   - permission: zessentials.vault.size.player
@@ -106,6 +135,7 @@ vault-permissions:
   - permission: zessentials.vault.size.admin
     slots: 500
 ```
+
 - Fixed title message placeholders
 - Fixed somes messages [#47](https://github.com/Maxlego08/zEssentials/issues/47)
 
@@ -119,7 +149,10 @@ vault-permissions:
 - Fixed CMI convert with invalid location for homes
 - Fixed the bug that allowed adding items in the mailbox even if the module is disabled.
 - Fixed command /rules who are not using the correct module.
-- Added the module `WorldEdit`, module allows players to have access to a **player worldedit**. They will be able to use the commands like `/pw set`, `/pw cut` to place or break blocks. Each block placed must be paid, by default 5$ per blocks. You can configure the item worldedit, the number of blocks that the player can change at the same time and many other things, more information [here](https://zessentials.groupez.dev/modules/worldedit).
+- Added the module `WorldEdit`, module allows players to have access to a **player worldedit**. They will be able to use
+  the commands like `/pw set`, `/pw cut` to place or break blocks. Each block placed must be paid, by default 5$ per
+  blocks. You can configure the item worldedit, the number of blocks that the player can change at the same time and
+  many other things, more information [here](https://zessentials.groupez.dev/modules/worldedit).
 
 # 1.0.0.4
 
@@ -128,28 +161,34 @@ vault-permissions:
 - Added the command `/ess convert CoinsEngine`, allows converting the database from CoinsEngine to zEssentials.
 - Items in player vault will be displayed as a single item.
 - Fixed night vision messages
-- Fixed duplicate lines [#33](https://github.com/Maxlego08/zEssentials/issues/33) 
+- Fixed duplicate lines [#33](https://github.com/Maxlego08/zEssentials/issues/33)
 - Upgrade to [Sarah](https://github.com/Maxlego08/Sarah/) version 1.10
-- Fixed issue where savings by default could not be removed [#34](https://github.com/Maxlego08/zEssentials/issues/34). You must update your configuration like this
-Before
+- Fixed issue where savings by default could not be removed [#34](https://github.com/Maxlego08/zEssentials/issues/34).
+  You must update your configuration like this
+  Before
+
 ````yaml
 economies:
   money:
     display-name: Money
 ````
+
 After
+
 ````yaml
 economies:
   - name: money
     display-name: Money
 ````
+
 - Fixed permissions for night vision and vault commands
 - Added command ``/sudo`` [#36](https://github.com/Maxlego08/zEssentials/issues/36)
 
 # 1.0.0.3
 
 - Change ``AsyncPlayerPreLoginEvent`` to ``PlayerLoginEvent``
-- Change /tp command for adding coordinate. You have now ``/tp <x> <y> <z> <yaw> <pitch>`` and ``/tp <player> <x> <y> <z> <yaw> <pitch>``
+- Change /tp command for adding coordinate. You have now ``/tp <x> <y> <z> <yaw> <pitch>`` and
+  ``/tp <player> <x> <y> <z> <yaw> <pitch>``
 - Fixed various messages in multiple languages
 - Fixed text hologram default text [#25](https://github.com/Maxlego08/zEssentials/issues/25)
 - Added command ``/nightvision`` (`/nv`), Provides a night vision effect
@@ -167,4 +206,4 @@ economies:
 # 1.0.0.1
 
 - Added the command `/ess convert CMI`, allows converting the database from CMI to zEssentials.
-Documentation: https://zessentials.groupez.dev/getting-started/convert#cmi
+  Documentation: https://zessentials.groupez.dev/getting-started/convert#cmi